Changes primary copy method and removes file dependency
Changes the copying method of the directories from xcopy to robocopy. This significantly reduces the LOC, and removes the dependency of an external file to list files/directories to exclude.

Minor changes include cleaning up the output, restructuring the logic to copy browser bookmarks, and changes to comments to reflect changes.